--- src/CMShareVideo/index.js
@@ -1,2 +1,0 @@
-export CMShareVideoEdit from './CMShareVideoEdit';
-export CMShareVideoView from './CMShareVideoView';
