--- src/CMShareVideo/players/index.js
@@ -1,6 +1,0 @@
-import YouTubeEmbed from './YouTubeEmbed';
-import VimeoEmbed from './VimeoEmbed';
-import CMShareEmbed from './CMShareEmbed';
-
-
-export default { youtube: YouTubeEmbed, cmsharevideo: CMShareEmbed, vimeo: VimeoEmbed };
