+++ src/NextCloud/index.js
@@ -1,0 +1,2 @@
+export NextCloudVideoEdit from './NextCloudVideoEdit';
+export NextCloudVideoView from './NextCloudVideoView';
