+++ src/NextCloud/players/index.js
@@ -1,0 +1,9 @@
+import YouTubeEmbed from './YouTubeEmbed';
+import VimeoEmbed from './VimeoEmbed';
+import NextCloudEmbed from './NextCloudEmbed';
+
+export default {
+  youtu: YouTubeEmbed,
+  nextCloud: NextCloudEmbed,
+  vimeo: VimeoEmbed,
+};
